requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.9.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.3.3
cycler==0.12.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.4.0
executing==1.2.0
fastapi==0.120.1
fastjsonschema==2.18.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
fqdn==1.5.1
h11==0.16.0
httptools==0.7.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.5.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.9
MarkupSafe==2.1.3
matplotlib==3.10.6
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==2.3.2
overrides==7.4.0
packaging==23.1
pandas==2.3.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==11.3.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.16.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.2.1
python-json-logger==2.0.7
pytz==2025.2
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.48.0
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.38.0
virtualenv==20.34.0
watchfiles==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.8
